requirements.txt
PyQt5
pyqt5-tools
bcrypt
Pillow
loguru